Make the platform struct not dependant on build flag (#103)
